Decide which network interface to monitor lazily
When awesome starts, network manager has yet to start, and hence
no physical network interfaces have IP addresses, and hence no network
interface is monitored via awesome's tray.

The solution I went for was to check which network interface to monitor
on every refresh of the widget. This means when the network interface
comes up, the monitoring will detect it on the next refresh.