Remove "--extended-regexp" from GREP_OPTIONS and alias grep=egrep
The problem with having "--extended-regexp" in GREP_OPTIONS is that
scripts read environment variables, and hence use grep with extended
regexp. This breaks many scripts.

Fortunately, scripts (i.e. non-interactive shells) don't expand aliases!

"Aliases are not expanded when the shell is not interactive, unless the
expand_aliases shell option is set using shopt" -- man bash

Thus the solution is to alias grep to egrep.