emacs: remove explain-pause-mode
Having explain-pause-mode enabled broke using with-editor to write
commit messages when committing to a repo over Tramp.

The process filter defined by with-editor (in
 https://github.com/magit/with-editor/blob/7c512887c6d69864fb600d32fb92857c51babcff/with-editor.el#L576
 ) wasn't being executed.

https://github.com/lastquestion/explain-pause-mode/issues/44 looks
related, but the fix
https://github.com/lastquestion/explain-pause-mode/pull/45/commits/ba7ad453697f024d4150b792023b4efa0f23d606
did not seem to resolve the issue.

I'll disable explain-pause-mode until I have time to raise a bug
upstream with a reproducible test case.